Adding custom multi-healthchecks
Adding the ability to add as many custom healthcheck scripts as required for this container, and removing the Docker based one.

Each script in /healthchecks.d/ will be looped through and run individually in sequence, and if any of them exit 1 then the container PID 1 is killed which should restart the container based on the Docker restart policy.